+++ modelrack4j-core/src/test/java/io/github/maxtrezzi/modelrack4j/testing/FakeAbsentProviderFactory.java
@@ -31,7 +31,7 @@ public final class FakeAbsentProviderFactory extends FakeProviderFactory {
     }
 
     @Override
-    protected boolean supportsModeration() {
+    public boolean supportsModeration() {
         return false;
     }
 }