+++ modelrack4j-core/src/test/java/io/github/maxtrezzi/modelrack4j/testing/FakeLocalProviderFactory.java
@@ -31,7 +31,7 @@ public final class FakeLocalProviderFactory extends FakeProviderFactory {
     }
 
     @Override
-    protected boolean supportsModeration() {
+    public boolean supportsModeration() {
         return true;
     }
 }